chore: add version.sh script to automate version bumping
The script automates the process of incrementing the version in package.json based on the specified release type (patch, minor, major). It also supports an optional --publish flag to build and publish the package to npm, along with committing and tagging the changes in git. This reduces manual errors and streamlines the versioning process.